Walking skeleton of a Compose service broker
A walking skeleton will give us a service broker which can be deployed
as a CloudFoundry application and registered as a broker, but with no
intention of actually provisioning service instances. This minimal
implementation serves the necessary endpoints for the CloudFoundry
Services API[1] and provides tests at the HTTP level to ensure proper
integration between the brokerapi HTTP handler and our implementation.

We make the broker serve the catalog endpoint, but the catalog provided
at this stage does not matter. It only serves to satisfy the Cloud Controller
during registration.

As this is a CloudFoundry application it should take its configuration
from environment variables. There was a question as to whether we should
provide the catalog via the environment. However, that would mean having to
use minified JSON, which is hard to read or would require us to convert
into a readable format, so a catalog JSON file was used instead for
simplicity.

[1] https://docs.cloudfoundry.org/services/api.html