Update Travis configuration and add test helpers
We want to run all except the tests against the real Compose API in
Travis. We run the integration tests using the fake Compose client and the rest
of the tests (of which there are currently none) as separate Travis scripts for two
reasons:

* it isn't easy to exclude a single set of tests using the focus regexp
* the tests against the real API share the same directory as the other integration
  tests because they are part of the same suite and share similar setup steps.
  This makes it impossible to exclude the tests based on the directory they
  are in.

We also add some test helpers for running the tests locally, because all of the
integration tests involve setting up environment variables and pulling credentials
from paas-pass.