Implement real API calls
This provides the first proper broker implementation which uses the
Compose client to interact with the Compose API.

Despite the `update` operation being implemented this functionality
is not technically supported, as we only have one plan available
and we do not have any user provided parameters. Upcoming tests
will assert that updates are not supported.

We make use of the `operation` query field from the Cloud Foundry
services API[1]. We set a string value on every provision, deprovision,
and update operation containing JSON. This JSON describes the operation
type and recipe ID which can be used to check on the status of the
operation using the Compose API.

However, a bug was encountered in which the Compose API returns a
401 when GETing the recipe for a deleted deployment. Compose have
recognised this and have it tracked as a bug. The workaround for
now is to assume delete operations were successful.

We return a single credentials string on bind operations as this
is an established convention and some buildpacks are configured
to use such URIs and place them in a DATABASE_URL environment
variable automatically.

[1] https://docs.cloudfoundry.org/services/api.html#polling