Add tests for the Compose API
These tests use a Compose client and the real Compose API to provide
a full end-to-end integration test. We have used the httptest library
instead of actually compiling the broker and running real HTTP requests.
This was because it is much simpler and was more consistent with the
existing integration tests. There is some duplication in how we construct
HTTP requests (those that would normally come from Cloud Controller),
but this is deliberate to aid readability.