DB instance prefix
Currently, a DB instance name managed by the broker is a plain UUID. That makes it harder for humans to recognise in Compose UI which instances are automatically provisioned by the broker. To make things a little bit easier this commit implements a DB prefix that will help to recognise which DBs are broker made.