Combine integration test suites.
Combining these avoids duplicating much of the setup code. This uses
Ginkgo's Skip function[1] to disable the tests that hit the real compose
API when needed.

As part of this refactoring, I've also reduced the use of global
variables in the tests. This results in the context for tests being
defined much closer to the individual tests themselves, which helps
readability.

[1]http://onsi.github.io/ginkgo/#pending-specs