Move lookup of account and cluster ID into broker.New
This removes this logic from main, and means that it can be tested more
easily. It also means that this is covered by the integration tests.

This has required extending the compose.Client interface to include
these extra methods. We were previously relying on a concrete type to do
these lookups.