Don't return ErrInstanceDoesNotExist for all errors
We were previously returning ErrInstanceDoesNotExist for any error
returned from the call to get the compose Deployment by name. This could
lead to erroneously reporting to CloudController that things are gone
when they aren't, and potentially lead to orphaned deployments.

As part of fixing this, I've switched to use the upstream client's
GetDeploymentByName() method which was recently added.