Add docker run commands into local tests script
This commit adds the running of the docker containers for local testing
against a real ElasticSearch and MongoDB. It also cleans them up when
complete.

This is a better solution than mocking the databases and creates an easy
to use interface for the tests.

An additioal Travis testrunner script has been added, this does not use
the docker containers, as both ElasticSearch and Mongo are available for
use.