Use the example catalog for integration tests
There was dupication of the json catalog data:

* There is an example catalog in used as a quick "getting started" reference
* There was a big string version in the integration tests helper

The duplication is not ideal as it means keeping two sets of the same
data up to date when adding new plans/services and also means that the
example catalog is not tested/validated.

This change makes the integration tests use the catalog at
examples/catalog.json so it is all in one place.