Switch to managaing dependencies with dep
Godep was proving difficult to manage when updating dependencies. This
switching to managing dependencies using the `dep` tool.

As part of this migration we update all vendored dependencies that can
be safely upgraded, add a redis client library and update the manifest
to set the GOPACKAGENAME to the fully qualified package name so that the
go buildpack can take advantage of Go's native vendoring.