Simplifly DBEngine interface
The DBEngine interface exposed database Open, Close and
ParseConnectionString methods that should really have been internal to
the engine.

This has now been simplified to:

```
type DBEngine interface {
	GenerateCredentials(instanceID, bindingID string) (*Credentials, error)
	RevokeCredentials(instanceID, bindingID string) error
}
```

The previous test for elasticsearch named "should successfully
ParseConnectionString() with more than one host" did not have an
assertion for checking the Port returned in the credentials .... during
refactoring we
added an assertion for the port in the new tests which surfaced the
an issue with the port returned in the credentials info and required a
workaround similar to the mongo engine (manually parsing out the first
port). A FIXME was left in the code since this is not ideal but the best
fix (changing the format of the binding crednetials struct to support
multiple host addrs) would be a disruptive change.