[WIP] Import paas-gocomposeapi fork with CacheMode
WIP because once the paas-gocomposeapi branch is merged to master we
will want to stop specifying the branch.

We need the ability to supply an extra `cache_mode` parameter to
Compose in order to support Redis services with CACHE enabled. We have
added this parameter to a fork of the official gocomposeapi library,
until such time as the official library has this functionality.

This commit vendors in our fork, by adding an `[[override]]` block to
`Gopkg.toml` and then running `dep ensure` to apply the override.