Set the authentication source when connecting to a MongoDB cluster
We start to parse the authSource parameter passed in the MongoDB connection
string. We use the authentication source to connect to the MongoDB cluster to
create/delete users and we send back this parameter for new service bindings.

Our authentication against an existing deployment stopped working at one point
and it turned out we are using the wrong authentication source (database).
Compose passes back an authSource paramater in the connection string which we
didn't use. We suspect the authentication source was originally the same as the
database that we connected to.

Currently a connection string looks like this:
mongodb://user:password@test-c00.2.compose.direct:17445/compose?authSource=admin&ssl=true.
'compose' is the database name we connect to, but we have to use the 'admin'
database as the authentication source. The mgo library by default falls back to
the connection database if the auth source is empty.

The compose documentation also says: "Users and authentication are handled by
the admin database". See
https://help.compose.com/docs/mongodb-connecting-to#section-connection-string-parameters.