MongoDB engine returns multiple portals
Add support for multiple portals in the mongodb engine.

We update the URI to support the multiple-endpoint format
documented by mongodb[1]. We accept that there might be
clients that do not support this format.

We replace the `host` and `port` with a list of strings in
`hosts` with the format `host:port`.

This might be a breaking change for some tenants if they
rely on the `host` and `port` or in having a simple host
definition in `uri`. If the tenant uses `uri` with a modern
mongodb client, it shall work OK.

[1] https://docs.mongodb.com/manual/reference/connection-string/

Signed-off-by: Andrew Bruce <me@andrewbruce.net>