ElasticSearch engine returns multiple portals
Add support for multiple portals in the elasticsearch engine.

We add a `uri` struct with the list of endpoints. The original
`uri` will only return the first endpoint returned by compose.

We replace the `host` and `port` with a list of strings in
`hosts` with the format `host:port`.

This might be a breaking change for some tenants if they
rely on the `host` and `port`. The shall be OK if they use the
simple host definition in `uri`. They can use all the endpoints
by using `uris`

Signed-off-by: Hector Rivas Gandara <hector.rivas.gandara@digital.cabinet-office.gov.uk>