Use a different creds struct per dbengine
We want to enrich the returned credentials for each dbengine
to include multiple endpoints (aka portals).

But each dbengine handle the credentials and the URIs for multiple
endpoints differently: In mongodb recommends a URI with a comma
separated of host:port[1] , ElasticSearch a list of URLs, etc.

Also, some attributes make sense in some engines (e.g. AuthSource and
name on mongoDB, but not on ES)

Because that, we create a definition of the credentials struct for
each DB engine, so we can evolve them independently.

[1] https://docs.mongodb.com/manual/reference/connection-string/

Signed-off-by: Hector Rivas Gandara <hector.rivas.gandara@digital.cabinet-office.gov.uk>