Restore-from-latest-backup feature
This adds support for creating new service instances from the latest
backup of an existing service instance in the same space. This is
tested for MongoDB and Elasticsearch, but not for Redis as we intend
to phase that out.

We've copied the approach used in the RDS Broker. The tenant will
provide a `restore_from_latest_snapshot_of` parameter containing the
GUID of an existing service instance. The broker will create a new
Compose deployment from the newest restorable backup of that existing
deployment.

Notably, we have started to store the GUID of the Cloud Foundry space
which owns a service instance in the Customer Billing Code of the
Compose Deployment.