Use "default" database name for all MongoDB service instances
Originally we generated a unique database name ("db_" + CF service UUID) for all service instances.

This makes it hard to restore a service from backup, as the backup contains the original database name which would be different from the expected name for the new instance. We would need to handle these exception which would cause confusion or we would need to rename the database which would be a slow process (there is no rename, so you can only copy it).

We decided to make all databases called "default" as there is no reason to have unique database names at all.

To preserve the backward compatibility the bind/unbind functions will list the existing databases and use the existing database name if there is one.